ember is faster than react when updating, but slower on initial render
