+++ frameworks/keyed/ember/app/templates/components/bs-button.hbs
@@ -1,0 +1,8 @@
+<button
+  type="button"
+  class="btn btn-primary btn-block" id={{@id}}
+  {{action @onClick}}>
+
+  {{yield}}
+
+</button>
