+++ frameworks/keyed/ember/app/templates/components/data-rows.hbs
@@ -1,17 +1,17 @@
 <table class="table table-hover table-striped test-data">
     <tbody>
-    {{#each rowData key='identifier' as |item|}}
+    {{#each @data key='identifier' as |item|}}
 
-        <tr class={{if (eq item.identifier selected) 'danger' 'x'}}>
-            <td class="col-md-1">{{item.identifier}}</td>
+        <tr class={{if (eq item.id @selected) 'danger' 'x'}}>
+            <td class="col-md-1">{{item.id}}</td>
 
             <td class="col-md-4">
-                <a {{action onSelect item.identifier}}>{{item.label}}</a>
+                <a {{action @onSelect item.id}}>{{item.label}}</a>
             </td>
 
             <td class="col-md-1">
 
-              <a {{action onRemove item.identifier}}>
+              <a {{action @onRemove item.id}}>
                 <span class="glyphicon glyphicon-remove" aria-hidden="true"></span>
               </a></td>
