+++ frameworks/keyed/ember/app/utils/benchmark-helpers.js
@@ -10,14 +10,45 @@ const updateData = (data, mod = 10) => {
     return newData
 }
 
+export function measure(name, fnToMeasure) {
+  const startTime = performance.now();
+
+  fnToMeasure();
+
+  const endTime = performance.now();
+
+  const duration = endTime - startTime;
+
+  console.log(`${name} took ${duration}`);
+}
+
 export const buildData = (id, count = 1000) => {
-    var adjectives = ["pretty", "large", "big", "small", "tall", "short", "long", "handsome", "plain", "quaint", "clean", "elegant", "easy", "angry", "crazy", "helpful", "mushy", "odd", "unsightly", "adorable", "important", "inexpensive", "cheap", "expensive", "fancy"];
-    var colours = ["red", "yellow", "blue", "green", "pink", "brown", "purple", "brown", "white", "black", "orange"];
-    var nouns = ["table", "chair", "house", "bbq", "desk", "car", "pony", "cookie", "sandwich", "burger", "pizza", "mouse", "keyboard"];
+    var adjectives = [
+      "pretty", "large", "big", "small", "tall",
+      "short", "long", "handsome", "plain", "quaint",
+      "clean", "elegant", "easy", "angry", "crazy",
+      "helpful", "mushy", "odd", "unsightly", "adorable",
+      "important", "inexpensive", "cheap", "expensive", "fancy"];
+
+    var colours = [
+      "red", "yellow", "blue", "green", "pink", "brown", "purple",
+      "brown", "white", "black", "orange"];
+
+    var nouns = [
+      "table", "chair", "house", "bbq", "desk", "car", "pony", "cookie",
+      "sandwich", "burger", "pizza", "mouse", "keyboard"];
+
     var data = [];
 
     for (var i = 0; i < count; i++)
-        data.push({ id: id++, label: adjectives[_random(adjectives.length)] + " " + colours[_random(colours.length)] + " " + nouns[_random(nouns.length)] });
+        data.push({
+          id: id++,
+          label: adjectives[_random(adjectives.length)]
+            + " "
+            + colours[_random(colours.length)]
+            + " "
+            + nouns[_random(nouns.length)]
+        });
 
     return { data, id };
 }
