+++ frameworks/keyed/ember/config/environment.js
@@ -6,7 +6,7 @@ module.exports = function(environment) {
     'ember-resolver': {
       features: {
         // EMBER_RESOLVER_MODULE_UNIFICATION: true,
-        EMBER_GLIMMER_ANGLE_BRACKET_INVOCATION: true,
+        // EMBER_GLIMMER_ANGLE_BRACKET_INVOCATION: true,
       },
     },
     modulePrefix: 'ember-temp',
@@ -18,8 +18,8 @@ module.exports = function(environment) {
         // Here you can enable experimental features on an ember canary build
         // e.g. 'with-controller': true
 
-        'ember-glimmer-angle-bracket-invocation': true,
-        EMBER_GLIMMER_ANGLE_BRACKET_INVOCATION: true,
+        // 'ember-glimmer-angle-bracket-invocation': true,
+        // EMBER_GLIMMER_ANGLE_BRACKET_INVOCATION: true,
       },
       EXTEND_PROTOTYPES: {
         // Prevent Ember Data from overriding Date.parse.
